Throw exception for error in fetch-hosted-zone-id
(breaking) Throw exceptions for errors in fetch-hosted-zone-id
instead of returning the anomaly response from aws-api. This
is more appropriate for how the function is typically used.
Its result is often fed directly into data structures,
so callers would have to provide their logic to throw exceptions
if we did not.