Add :cloudformation-client, deprecate :client
(deprecation) Rename `:client` config option
to `:cloudformation-client` for `change-set`, `stack`, and
`stack-properties` components.
`:client` is still accepted, but is deprecated.
Salmon is normally used in systems that call multiple AWS services.
With this change, you can construct a single opts map with keys like
`:cloudformation-client` and `:s3-client` that can be passed
directly to various components.